[#285] ELSE keyword in CASE statements now work
The main fix is to src/m_templates/tmpl_print_expression.ctemplate to set the `default_value` variable
correctly. Due to presumably a typo, it was set to `cur_plan->v.operand[0]` (which houses the value
part of the CASE) when instead it should have been set to `cur_plan->v.operand[1]` (the optional else
part of the CASE).

In addition, the optional else part of the case statement was not qualified in qualify_statement.c
and that is now done. This fixes the case where a column name is used in the ELSE part of the case.

New bats test test_case_statement covers the cases described in the #285 issue description on gitlab.

While at this, a few template generation files were changed to use full names (instead of abbreviated
names) for M functions. This caused a few reference file fixes in the test_where_in bats test.