Fix failing tests previously masked by ommission of diff command in verify_output helper function
A call to `diff` was accidentally removed from the verify_output helper function during a merge conflict resolution, causing several test failures to be masked.

These failures are fixed by this MR, with the exception of those in test_cancel_request.bats. These have proven difficult to resolve and so have been skipped and will be addressed in a separate MR.

The changes made include an updated outref and the addition of support for a secondary option to the start_rocto helper function for compatibility with a recent test change. Also, read_cancel_request was updated to use integers in its error messages instead of strings, as it was exempted from parallel changes to error message types.

Additionally, a few non-portable/variable size type declarations have been changed to fixed size.