[#177] Corrected test_cancel_request.bats failures by removing C SIGUSR1 handler
The recent addition of a SIGUSR1 handler to rocto.c led to test failures in test_cancel_request.bats. This is assumed to be the result of a lack of support for SIGUSR1 forwarding from C code to YottaDB. Accordingly, the offending signal handling code has been commented out and comments added explaining the reasons.

Additionally, the expect script used in test_cancel_request.bats has been renamed, an additional test added, and an additional option added to the verify_output helper function.