[#222] Optimize queries that have a mix of INNER and OUTER JOINs in the same query
As part of a recent commit (SHA 2cd272fd11d070a258a5ca78fc1d966cfcec4514), a call to
lp_optimize_where_multi_equal_ands() was skipped if the query had at least one outer join.
This was needed because one cannot move the ON condition of the OUTER JOIN into the WHERE
clause like one can for INNER JOINs.

But we were maintaining the join condition only for OUTER JOINs. This meant that for a query
which has an INNER JOIN and an OUTER JOIN, we will move the ON condition corresponding to the
INNER JOIN into the WHERE clause but keep the ON condition corresponding to the OUTER JOIN
as part of the outer join (table_join->outer_join_condition). We would invoke the optimization
function lp_optimize_where_multi_equal_ands() on the outer join condition but not on the inner
join condition since that got moved into the WHERE clause and we skipped the WHERE clause
invocation.

To address this case, we now maintain the join condition even for INNER JOINs. The field
outer_join_condition has therefore been renamed as join_on_condition. If we find that there
are no outer join usages in the entire query, then we move the ON condition of the INNER JOIN
into the WHERE condition. Otherwise we keep it in the INNER JOIN join_on_condition field.
This way when lp_optimize_where_multi_equal_ands() gets invoked for the INNER JOIN join_on_condition
any keys that can be fixed end up getting fixed in turn optimizing queries that have a mix of
INNER and OUTER JOIN usages.

A new template function tmpl_tablejoin_innerjoin() generates the ON condition of the INNER JOIN
if it is non-NULL.

An existing reference file (TCD07.ref) was modified to reflect the new output (minor change).

With this optimization, a SquirrelSQL catalog query that used to take minutes to run now takes a
few seconds to run.