[#299] ORDER BY when used with LIMIT should return correct results
The main fix is to tmpl_tablejoin.ctemplate to NOT generate M code to implement
the LIMIT in case ORDER BY has been specified. This is because we want to do the
LIMIT on the output generated by ORDER BY which is only available later. In addition,
tmpl_physical_plan.ctemplate is modified to generate M code to implement the LIMIT
qualifier in case ORDER BY has been specified.

The existing genrandomqueries.m M file has been renamed to be genrandomqueries.m and
enhanced to also generate queries with ORDER BY and LIMIT thereby testing this fix.

An existing reference file TWO04.ref had to be modified to reflect the slightly modified
generated M code.