[#257] Make a couple variables non-static to avoid double-free/corruption issues
`data_row_parms` is a variable used in `handle_query_response.c`. This was defined as `static`.
This meant that one invocation of the function could malloc some memory and assign it to this
global variable. The second invocation of the same function would then use the same memory
instead of allocating memory afresh.

Recently, as part of the #257 efforts to avoid memory leaks, a `free(data_row_parms)` was done
in `handle_query_response.c`. But the variable was not set to NULL. This meant that in another
invocation of the same function we would notice the `static` variable to have a non-NULL value
and try to do a `free(data_row_parms)` again resulting in a double-free.

To fix this, one needs to also set `data_row_parms = NULL` immediately after the `free(data_row_parms)`.

A better fix would be to remove the `static` type from the variable. That makes it non-global and
therefore every function invocation would not inherit leftover memory from a prior invocation.
This is the approach taken in this commit. That avoids us having to do a `data_row_parms = NULL`
right after the `free(data_row_parms)` since we are going to return from this function invocation
anyways.

Additionally, one code path that was not doing a `free(data_row_parms)` now does it. This was okay
previously since the leftover memory would be reused but now it will not be so we need to ensure
all code paths free up any memory they allocate.