[#321] LIMIT returns more rows than requested if there is a WHERE clause with OR usages
The main fix for this is to tmpl_tablejoin.ctemplate to also generate M code for checking
the LIMIT clause in case a key is fixed (key->value is non-NULL). Previously we were
doing this only in case key->value is NULL.

While at this, it was noticed that we were unnecessarily generating M code to use more
subscripts than necessary in tmpl_duplication_check.ctemplate and that has been fixed
so we reduce the chances of a MAXNRSUBSCRIPTS error. The test_composite_key bats test
has been enhanced to test this. While coming up with this test case, fixed
tmpl_physical_plan.ctemplate to not generate multiple identical M lines (to generate
the xref) in case of a composite key with multiple key columns (all of which have the
same cross_reference_output_key).

Existing reference files were fixed to account for the revised generated M code.

The specific test identified in the description of #321 in gitlab has been added to the
test_limit bats test. Took this opportunity to also add the specific test in the description
of #299 to the test_limit bats test.

genrandomqueries.m has been enhanced to test for #321 and #299 in a variety of ways. The WHERE
clause is now randomly generated with a variety of boolean expressions.

genrandomqueries.m and test_helpers.bash.in have been together enhanced to also do testing
in case LIMIT has been randomly generated by just counting the # of rows from psql vs octo
instead of comparing the actual output.

Note that an occasional randomly generated query could fail due to #324 (fixed in a separate
upcoming commit).