[#324] WHERE clause with <, >, <=, >= does not work correctly for STRING types
The primary fixes are to tmpl_print_expression.ctemplate.  It was using the ]] operator instead of
the ] operator in M for string comparison. Not clear why the former was chosen. Using ] instead of
]] addressed the incorrectness issues with string comparisons. In addition, <= and >= were also
not correctly implemented for string types. They were using '> or '< operators in M which will not
yield the right results for string comparisons. That is also fixed now to use the ] or '] operator
as appropriate (by swapping the operands if needed).

genrandomqueries.m (which is what generated a query that exposed the issue #324) has been enhanced
to use <= and >=.

This commit also enhances tmpl_physical_plan.ctemplate to not generate unnecessary M code to check
and generate xrefs. This also resulted in changes to the reference file TCK03.ref.

In addition, this commit tries to address an occasional failure of the test_where_optimization bats
subtest TW016 where a new line between two consecutive queries specified in the input to octo
occasionally goes missing only in the pipeline resulting in the Octo parser seeing two queries
one after the other with just a `;` separating them (no other white space) resulting in the
first keyword `select` being interpreted as `elect` by the parser and issuing an error. The octo
parser should still work in case there is no space separating two queries as long as a `;` separates
them. https://gitlab.com/YottaDB/DBMS/YDBOcto/issues/325 has been created to track this. But we
want this test to pass in the meantime so a workaround for now is to copy all the queries into a
file initially (instead of letting the shell create a HEREDOC) and use this as the input to octo
hoping this would not let the newline go missing.