Updated find_open_port helper function to account for processes running as different users
The find_open_port helper function previously used lsof to check for open ports. However, this utility is limited to files opened by the user running it. To allow this function to account for processes run by other users, this command was replaced by netstat.

Also, several tests in test_psql_connection.bats failed due to port reuse when run while a rocto process was run parallely. This was resolved by storing the port returned by the start_rocto helper function and passing this port to the run_psql function, instead of the port assigned at the start of the test.