[#269] UNION does not work correctly if operand SELECT statements use LIMIT to limit the number of rows
For the below SQL query, the output key for both sub-queries (operands of the UNION) was the same
so the generated M code ended up using the same output key for the LIMIT in each sub-query.
This meant the output key count became 2 at the end of the first sub-query and after adding one
more record from the second sub-query the output key became 3 and so quit the second sub-query
instead of adding 3 rows from the second sub-query.

```sql
(select firstname from names limit 2) union (select lastname from names limit 3)
```

The fix is to not use the output key (which could be the same across the generated physical plans)
but instead use a local variable (`%ydboctoz("limit")`) for the check to know if the LIMIT has been
reached in each plan. This local variable is incremented whenever the output key is incremented in
each plan.

This fix worked for the above sql query but the below second sql query posed an issue.

```sql
select lastname from names where id = 1 or id = 2 limit 1
```

Because the above query effectively gets treated as a UNION of two queries, i.e. as follows

```sql
select lastname from names where id = 1 limit 1
```

AND

```sql
select lastname from names where id = 1 limit 1
```

But in this case, we do NOT want to keep the local variable as we want the total limit in this case
to be 1 (i.e. individual query limits should not be honored like in the previous example). Towards
this, a new `is_prev_plan_in_same_dnf()` function is implemented which will return TRUE if a plan
and its previous plan are both part of the same dnf (like in the second sqsl query above). This is
used in `tmpl_tablejoin.ctemplate` and `tmpl_tablejoin_body.ctemplate` to decide whether to generate
M code to use the output key or `%ydboctoz("limit")`.

The test_limit bats test has been enhanced to test for the first sql query above. It already tests the
second sql query above.