[#315] Changes to LIKE to fix the SQL standard, and addition of ILIKE and SIMILAR TO operators
Previously octo would convert `%` in to `.*` and `_` in to `.`, but
would not escape literal regex characters like '.', '*', '[', and other
meta characters like '|' could be enabled with a '\'. Now octo properly
escapes, or unescapes regex characters before passing them to the
engine.

Additionally the postgres standard spesifies ~~ (equivlent to like),
ILIKE (case insensitive like), and SIMILAR TO, which have been added to
the parser.