[#334] ORDER BY on a column with empty string values does not work correctly
The main fix is to tmpl_physical_plan.ctemplate. To start the first iteration of the loop
with the "" subscript (in case $DATA shows that node exists) in case of ascending order.
Previously we would do it unconditionally at the end because of the way the M code was written.
We still need to do it the old way in case of descending order.

A new TOB05 subtest in the test_order_by bats test verifies the new behavior.

Two existing subtests (T2003.ref and T2004.ref) also verify the same behavior. They had
incorrect reference files which changed with the code fix so the reference files are now
fixed to reflect the correct output order.

TWO04.ref needed a change to reflect the new M code.

test_composite_key.bats.in was changed to filter out the FOR loop due to the "order" subscript
that showed up now because of the rearrangement of the M line and the way the sed/grep commands
were done previously. It continues to not show up (as it is not important to this particular test)
after the changes to grep out any lines with `"order"` usage.