[#338] SELECT from sub-query now returns correct results if sub-query had constants in its select column list
The main fix is to `src/get_column_piece_number.c` where in the `column_STATEMENT` case, we were
going through some fancy logic to figure out the number of this column in the table. As part of a
recent commit, the field `column->column_number` was introduced and so this is readily usable
and avoids a lot of the complexity in this file. Now that the correct column # is being returned
(previously it used to always return 1 in case of a DERIVED_COLUMN from a sub-query) the right
pieces are extracted in the parent query resulting in the correct output.

A new TSC14 subtest in the test_select_columns bats test verifies this issue is fixed.
On a related note, various existing reference files in the test_select_columns test were
renamed to follow the new convention (TOO*.ref to TSC*.ref).