[#340] Fix incorrect <Unknown column> error for column names inherited from sub-query
As part of a prior commit (SHA f5d61fd7b0eb8e3e13f3824c3e29ed9cec169290, !299), a change
was done to pass an additional parameter to be used inside `match_column_in_table.c` and
`qualify_column_name.c` to decide whether strict column name validation is needed or not.
While the change to `qualify_column_name.c` was necessary, the change to `match_column_in_table.c`
was not. That change is what caused queries like `select a from (select 1 as a)` (described
in the issue description of #340) to incorrectly fail with an invalid column name error.
This is because the column is a `value_STATEMENT` (hardcoded value of 1 in the sub-query,
a numeric literal) but even in that case we can safely use the alias (`a`) in the sub-query
for checking if the column name `a` specified in the outer query is valid. The additional
parameter and associated code has now been removed.

In addition, some cleanup of `qualify_column_name.c` was done to match coding conventions.

A new `TSC15` subtest has been added to the `test_select_columns` bats test.