Added additional expect line to catch initial command output in case of latency
test_memory_usage.bats has been observed to fail in rare cases, with rocto issuing an incorrect password error. This was found to be due to the get_mem_usage.exp expect script sending the password prematurely when it takes psql an unusually long time to start.

To make this test more robust, the password prompt timeout was extended in the expect script. This should decrease or eliminate cases of the observed failure. If it recurs, an increased timeout may be called for.