Added expanded timeout in two expect scripts to wait for password prompt in case of latency
test_cancel_request.bats has been observed to fail two subtests under
some conditions, with rocto issuing an incorrect password error. This was found to be due to
the cancel_psql*.exp expect scripts sending the password prematurely
when psql takes an unusually long time to start.

To make these tests more robust, the timeout for expecting the password prompt was expanded. This should decrease or eliminate
cases of the observed failures. If they recurs, an additional timeout increase may be called for.