[#348] Fix issue where sub-query that uses column value from parent query generates incorrect results due to an incorrect cross-reference plan
As part of a rework of `src/m_templates/tmpl_tablejoin_body.ctemplate` done in a previous commit
(SHA 976d9e003226bbc1796be9b943d9960338e0364b) the case where a cross-reference plan happens
in a sub-query was incorrectly coded. This is now fixed by checking if it is a cross-reference plan
and if so not worry about checking `plan->stash_columns_in_keys` (which implies it is a sub-query).
This simple change fixes the #348 issue.

An existing test reference file TWI01.ref is also fixed to reflect the new (and correct) output.

A new TSS02 subtest has been added to the test_select_subquery bats test to additionally verify this fix.