Updated Rocto to use Error Macros, instead of manually sending them
Error messages in Rocto were previously declared and sent manually, which has been changed to use Macro calls instead.

This change also included disabling the sending of notification message to client during authentication. PSQL Clients do not expect notification messages during authentication, but Rocto previously sent messages during authentication causing failures

The underlying problem turned out to be that the placement of the rocto_session.sending_message = FALSE; line in rocto.c was disabling error messages from being propagated to the client during authentication. This placement was half-correct insofar as the client doesn't expect any Notification level messages to be sent here, but without an error message for a FATAL error the client just gets confused and gives up.

The fix was to move the above line higher in rocto.c to enable message sending, but to disable it temporarily in handle_password_message.c on successful authentication so that Rocto logs the authentication internally without issuing an unexpected Notification to the client.