[#358] Fix Assert failure in yyerror when `!;` is entered at OCTO prompt
This was a case where llocp->first_column and llocp->first_line was 0 but llocp->first_column was > 0.
In this case, `print_yyloc()` would have done the right thing. So the if check has been modified
to add a `||` clause.

A new TERR009 subtest in the test_errors bats test has been added to verify this fix.