Speed up test_sqllogic_select1 bats test run time in pipeline (from 300 seconds to 17 seconds)
Although this test has 1000 subtests, only 326 of them run currently due to open Octo issues.
This takes 308 seconds to run in the pipeline. Therefore it is almost 1 second per subtest.
The suspicion is that the mkdir done in each subtest is the slowing step and so we avoid
it by creating file names in the top level directory itself corresponding to each subtest
with a prefix that is unique to each subtest.

Also, the `test_sqllogic_select1` bats test now uses the `init_test` function like all other bats
tests. The `init_sqllogic_test` function is removed and the small part of that function which is
different from `init_test` has been folded into `init_test`. This way all other changes done to
`init_test` automatically get included for the `test_sqllogic_select1` test too.

Update: With this change, the `test_sqllogic_select` test which usually takes 300+ seconds to
        run in the pipeline ran in 17 seconds so clearly mkdir was the slowing step.