Speed up test_memory_usage test in make-ubuntu and make-tls-ubuntu pipeline jobs from 7 minutes to 1 second
The test_memory_usage test took 7+ minutes to run only on the make-ubuntu and make-tls-ubuntu pipeline
jobs. It ran in a few seconds in the make-centos and make-tls-centos pipeline. Turns out the output
that came to psql (through a rocto -vv command) was huge so psql invoked the paginator by default
(I guess it is the Unix command `more`). And that confused the expect script which was expecting a
`ydb=.` prompt but not seeing it since psql was displaying only a partial output and waiting for the user
to key in a space prompt to see more output and finally hit q to quit out of this and go back to the
psql prompt. Because the expect script was not ready for this, it ended up timing out. Since the timeout
was set at 60 seconds and there were 7 or so steps, this caused the test to run for 7 minutes before
timing out (but incorrectly passing). This is now fixed by setting the pager to off in psql.

And the issue about incorrectly passing even if the expect command times out is fixed by the test now
verifying all needed plans are generated.