[#257] Fixed bug in issuing ParameterStatus messages by adding missing brackets and [#309] corrected bug due to misplaced free in handle_describe
This commit corrects two bugs and cleans up an unused `else if` block.

[#257] The first bug occured at approximately line 446 in rocto.c due to omitted brackets for an `if` statement, which caused the surrounding loop to break prematurely. This resulted in ParameterStatus messages not being sent.

In turn, this led SquirrelSQL to issue a query that prefixed certain literals with an 'E' character, causing the parser to issue errors and tables/columns to not be loaded.

A test was added using the Go pq driver to validate that ParameterStatus messages are sent by rocto. This also included updating existing Go tests to use outrefs instead of merely ensuring rocto doesn't crash.

For reference, this bug was introduced by commit c9287855 and is observable with commit 6a5e7446. The latest commit without this bug was 3b21354f.

[#309] The second bug stemmed from a misplaced pair of YDB_FREE_BUFFER calls in handle_describe that led to an assert failure in parser.y where NULL == cursorID. These have been moved to the correct location, preventing this error.

For reference, this bug was introduced by commit 64c5c10.

This commit also includes revisions to test_seed_queries.bats.in so that its subtests now use outrefs instead of merely expecting rocto not to crash. The SQL queries in question were also moved to fixture files for maintainability.

Additionally, an empty 'else if' block was removed from rocto.c as it is unnecessary for correct functionality and introduced erroneously.

As a fix for a pipeline misfeature, instructions to install Go have been added to the .gitlab-ci.yml before_script configurations for both Ubuntu and CentOS pipelines.

A verify_output option was also added to remove hex strings to abstract addresses from outrefs.