[#363] FULL JOIN does not work in cases where ON condition has equality check of primary keys
The issue is best described using an example. Take the following query and its associated M plan.

```sql
      1 ;; This is a generated file; do not modify. Generated M code corresponds to below SQL query
      2 ;; ---------------------------------------------------------
      3 ;  select n1.id from names n1 full  join names n2 on  1 = 0 and n1.id = n2.id;
      4 ;; ---------------------------------------------------------
      5
      6 octoPlan0(cursorId)
      7     DO octoPlan1(cursorId)
      8     QUIT
      9
     10 octoPlan1(cursorId)
     11     NEW %ydboctoz KILL %ydboctocursor(cursorId,"keys",3,"","")
     12     SET %ydboctocursor(cursorId,"keys",0,"NAMES","ID")=""
     13     FOR  SET %ydboctocursor(cursorId,"keys",0,"NAMES","ID")=$ORDER(^names(%ydboctocursor(cursorId,"keys",0,"NAMES","ID"))) QUIT:(%ydboctocursor(cursorId,"keys",0,"NAMES","ID")="")  DO
     14     . SET %ydboctoz(1)=0
     15     . SET %ydboctocursor(cursorId,"keys",1,"NAMES","ID")=%ydboctocursor(cursorId,"keys",0,"NAMES","ID") IF ($DATA(^names(%ydboctocursor(cursorId,"keys",1,"NAMES","ID")))'=0) DO
     16     . . IF ((+$GET(%ydboctocursor(cursorId,"parameters",1))=+$GET(%ydboctocursor(cursorId,"parameters",2)))) DO
     17     . . . SET %ydboctoz(1)=1
     18     . . . SET %ydboctoz(1,%ydboctocursor(cursorId,"keys",1,"NAMES","ID"))=""
     19     . . . IF 1 DO
     20     . . . . SET:1&$INCREMENT(%ydboctocursor(cursorId,"keys",3,"","")) %ydboctoz=42
     21     . . . . SET:1 %ydboctocursor(cursorId,"keys",3,"","",%ydboctocursor(cursorId,"keys",3,"",""))=$GET(%ydboctocursor(cursorId,"keys",3,"","",%ydboctocursor(cursorId,"keys",3,"","")))_%ydboctocursor(cursorId,"keys",0,"NAMES","ID")
     22     . IF %ydboctoz(1)=0 DO
 ->  23     . . SET %ydboctoz(1,%ydboctocursor(cursorId,"keys",1,"NAMES","ID"))=""
     24     . . IF 1 DO
     25     . . . SET:1&$INCREMENT(%ydboctocursor(cursorId,"keys",3,"","")) %ydboctoz=42
     26     . . . SET:1 %ydboctocursor(cursorId,"keys",3,"","",%ydboctocursor(cursorId,"keys",3,"",""))=$GET(%ydboctocursor(cursorId,"keys",3,"","",%ydboctocursor(cursorId,"keys",3,"","")))_%ydboctocursor(cursorId,"keys",0,"NAMES","ID")
     27     SET %ydboctocursor(cursorId,"keys",1,"NAMES","ID")=""
     28     FOR  SET %ydboctocursor(cursorId,"keys",1,"NAMES","ID")=$ORDER(^names(%ydboctocursor(cursorId,"keys",1,"NAMES","ID"))) QUIT:(%ydboctocursor(cursorId,"keys",1,"NAMES","ID")="")  DO
     29     . IF '$DATA(%ydboctoz(1,%ydboctocursor(cursorId,"keys",1,"NAMES","ID"))) DO
     30     . . IF 1 DO
     31     . . . SET:1&$INCREMENT(%ydboctocursor(cursorId,"keys",3,"","")) %ydboctoz=42
     32     . . . SET:1 %ydboctocursor(cursorId,"keys",3,"","",%ydboctocursor(cursorId,"keys",3,"",""))=$GET(%ydboctocursor(cursorId,"keys",3,"","",%ydboctocursor(cursorId,"keys",3,"","")))_""
     33     QUIT
```

A FULL JOIN is implemented as a LEFT JOIN plus a RIGHT JOIN. Line 17 implements the LEFT JOIN side by saying
we found a match (one right table row) for the current left table row. Line 18 implements the RIGHT JOIN side
for the matched right table row by noting down the primary key of the right table which matched. That is all
good.

Later we come to line 22 where we check if the current left table row did not match ANY right table row
(i.e. %ydboctoz(1)=0). In that case, we generate line 23 which, just like line 18, implements the RIGHT JOIN
side for the matched right table row by noting down the primary key of the right table which matched. But in
this case, we did not have a match at all (that is the point of the IF check in line 22). And therefore, we
should not have generated line 23. Removing Line 23 will automatically fix the plan and return correct results.

This is done by the `if (!this_join_is_left_join || !plan->treat_key_as_null[left_table_id])` check that is
newly introduced in `src/m_templates/tmpl_tablejoin.ctemplate`. This handles the `NO_MATCHING_RIGHT_TABLE_ROWS`
code block case.

As part of this fix, I realized it is easier to keep all the inner/left/right join template code inline in
`src/m_templates/tmpl_tablejoin.ctemplate` instead of in separate template files (or else we would need to
pass a lot of parameters to them all of which are available as local variables already in the only caller
of those helper functions) so all of those files have been removed and folded inline in the only caller
which is `src/m_templates/tmpl_tablejoin.ctemplate`.

The `test_outer_join` bats test has been reworked to add tests for this issue. Additionally,
pre-existing tests there have been greatly enhanced by comparing Octo output for those queries
against Postgres and ensuring they are identical. Previous versions of the tests did not check the
actual output of those queries and so could pass even if they did not return accurate results.