Set LC_ALL to fixed value en_US.UTF-8 so calling environment does not cause test failures due to sort order differences
sort behaves differently when LC_ALL is set to C vs when it is set to en_US.UTF-8.

Depending on the actual input data, the sorted output might show up differently.
And the test might pass if LC_ALL is set appropriately in the caller environment.

To avoid such environment-specific failures, the bats test framework now sets the locale
to en_US.UTF-8 for all tests. A few reference files needed to be fixed to account for
this change.