[#383] Updated load_fixture helper to strip copyright and comments from SQL fixtures
SQL fixtures contain copyright information and occasional comments.
These were previously removed on a case-by-case basis in a number of
BATS tests. The load_fixture helper function has been revised to do this
cleanup automatically, sparing the developer from manually
copying/writing the same code for similar tests in the future.

As a part of this change, a number of BATS subtests and outrefs were
updated to use this revised helper function.