Enhanced `stop_rocto` helper to wait for the `rocto` process to die before returning
It was discovered that test_memory_usage.bats occasionally fails in the
pipeline due to a failed grep on rocto.log. However, review of the
artifacts showed that the target output was in fact there. This strongly
suggests that a timing issue led to the grep occurring before the target
output had been written to the log.

To prevent this in the future, the `stop_rocto` helper function was
enhanced to wait for the `rocto` process to die before returning. This
should resolve similar log population timing issues in this and other
tests.