[#446] Corrected SquirrelSQL column display regression caused by [#94]
The integration of the #94 changes in !453 resulted in missing column information from the "TABLE" dropdown menu under the "public" database in SquirrelSQL. This is now fixed such that SquirrelSQL provides column information from this dropdown.

There were two causes behind this regression:
        1. Omission of `BOOLEAN`s from the `ParseContext` as they are not treated as literal values (i.e. with `LITERAL` token)
        2. A missed edge case in the identification of Bind parameters in `handle_parse`

More specifically, the edge case in point 2 pertains to a heap-allocated array used to track whether a parameter is a literal vs Bind parameter is initialized. This array is initially allocated to a fixed size and resized on demand. However, when the number of Bind parameters is less
than the default size AND the total number  of parameters is greater than this default size, it won't get resized. This led to erroneous accesses of uninitialized memory leading to incorrect identification of parameters as Bind parameters (false positives), in turn leading to omitted
parameters from the total.

The fix for the omitted BOOLEANs was simply to increment `parse_context->total_parms` when they are encountered, resulting in all parameters being accounted for on the `parse_context`.

The fix for the missed edge case in `handle_parse` was to add a check for `(cur_parm <= parse_context.is_bind_parm_size)` to the beginning of the parameter value population loop in `handle_parse`.

Also, test_squirrel_sql_connect_queries.bats.in was revised to use the JDBC driver, including updated outrefs and the addition of a new `strip_sql_comments` helper function for use in `run_java`. With a change to `run_query.java`, these tests now randomly use the extended query protocol by setting the JDBC connection parameter `preferQueryMode` to `extended` in `run_query.java`.

Further, a number of new tests were added to test_squirrel_sql_connect_queries.bats.in to account for all known connection/column retrieval queries. This change revealed that two calls to `parse_literal_to_parameter` were unaccounted for when tracking the total number of query parameters, which led to missing return rows. This was corrected by making the following changes:
+ Moved the `parse_context->total_parms` incrementation to `parse_literal_to_parameter` as there should be one such incrementation per call to this function. This fixes the above issue and similar issues that may have arisen in the future by coupling these related operations.
+ Moved the `cursorId` string parameter onto the `ParseContext` to simplify a number of interfaces and improve maintainability
+ Revised a number of interfaces to match the above, along with calls using them in various files:
        - `parse_literal_to_parameter`
        - `parse_line`
        - `populate_data_type`
        - `yyerror`
        - `regex_specification`
        - `INVOKE_PARSE_LITERAL_TO_PARAMETER`
        - `INVOKE_REGEX_SPECIFICATION`

Additionally, some miscellaneous enhancements:
+ Revised `make_command_complete`, `test_command_complete`, and `parse_context` to use the `SqlStatementType` enum instead of a string to track command tag information. This improves performance, maintainability, and safety by allowing the removal of `strncmp` operations to check command tag values.
+ Added missing emission of `BindComplete` message to early return case in `handle_bind` for PostgreSQL wire protocol compliance
+ Revised error case for `Parse` messages in `rocto_main_loop` to match other cases - this was a previous omission that is now corrected
+ Revised `run_java` and crosscheck helper functions for compatibility with generated queries
+ Moved TJC003 to a fixture and deleted TJC000 as redundant since it this query is the same as TSSCQ003
+ Added new `noconfig` option to `verify_output` helper function for removing Octo's config loading INFO message while retaining other INFO messages
+ Removed duplicate TJC007 test
+ Updated `run_query_generator` helper function to apply `usejdbc` option (was previously discarded)
+ Revised stop_rocto helper to wait for child processes to die in addition to waiting for the listener process to die
+ Renamed various query files in test_jdbc_connection.bats.in from TQG to TJC
+ Revised strip_sql_comments to remove .sql extension before if check for compatibility with calls from run_query_in_octo_and_postgres_and_crosscheck