[#415] Added instructions on using sample data to README.md
Previously, the Octo quick start documentation did not include reference to any sample data should users desire to do manual testing of Octo. This is now provided in README.md in the form of a new Test with dummy data section in the Quickstart section.

This new section provides information on how to load the Northwind and Customers data sets, as these should be sufficient for manual testing purposes.