[#347] Added unit test to validate asserts are removed in release builds; Rework octo_init.c for simplicity/correctness
Issue #347 was resolved with !528, however that MR did not contain any tests to validate this solution. A test to confirm the absence of asserts has been added with this change. Specifically, a new unit test was added that asserts failure and so will failure if asserts are present and succeed if they are not.

Additionally, `octo_init.c` was revised for readability and correctness by renaming variables to be more explicit, simplifying buffer population, and revising the handling of `$ZROUTINES` values in the loop that populates `config->tmp_dir`. Changes included:
+ Adding a new `SET_CONFIG_VARIABLE_NAME_AND_RETURN_ON_ERROR` for use in `populate_global_names`, reducing code duplication and adding error checking where there was none previously.
+ Adding type casts to various uncasted pointers returned from calls to `malloc`
+ Renaming ambiguous variable names to explicit ones, e.g. `c` -> `config_path_len`
+ Revising the retrieval of the value of `$zroutines` to use a loop to reduce code duplication, including the addition of a missing error check on a `ydb_get_s` return value
+ Adding a missing `break` to the `config->tmp_dir` population loop to handle the case when `$zroutines` contains empty parentheses for the first path
+ `populate_global_names` was revised to return an error value, which is now handled where none was issued or handled previously
+ Adding new MERGE_CONFIG_PATH macro to reduce code duplication for merging various paths from configuration files
+ Revising members of the `GlobalNames` struct in config.h to be statically allocated to fixed sizes to reduce `malloc` and improve performance, as the max size of these members is both known and small
+ Adding a new TOZ014 test case that tests logic for an edge case in `octo_init.c` where ZRO_INIT_ALLOC is the length of `$zroutines` plus the `octo_zroutines` config file setting
+ Removed the obsolete disabling of `test_port_option` on CentOS as this test is no longer slow
+ Renamed CMake project from 'OctoSQL' to 'YDBOcto' and updated various `#define`s and files accordingly

Finally, an `ERR_BAD_CONFIG` config call was reduced in severity from `FATAL` to `ERROR` as this is a known/expected error and should be handled, rather than crashing and generating a corefile. Similarly, various `FATAL` errors for handling errors when reading from a configuration were converted to `ERROR`s.