[#255] Added row limiting support for Extended Query exchanges
Per #255, Rocto previously did not respect result row limits or issue PortalSuspended messages. This is now corrected such that Rocto limits the number of rows returned in response to an Execute message with a row limit specification and, if not all rows are returned, returns a PortalSuspended message to the client.

Implementing this feature entailed some substantial revisions to `handle_execute` and `handle_query_response`, including some interface changes. Specifically:
1. `DataRow` generation and emission was moved from `handle_query_response` into a separate `send_result_rows` function to avoid cursor generation when a portal is suspended.
2. Now both `handle_execute` and `handle_query_response` each call send_result_rows, which handle all the row generation logic.
3. The interface to `handle_execute` was modified to take a cursorId parameter so that the cursor of a suspended portal can be reused to retrieve remaining rows when a limit is specified, rather than re-running the query on a new cursor.
4. The logic of `handle_execute` and `handle_query_response` was revised to limit the number of rows when specified, or to return all rows for simple queries or when no limit is specified.
5. A new rows_remaining database node `(^ydboctosession(session_id,"bound",portal_name,"rows_remaining"))` for each new portal was added to track how many rows remain to be sent to the client when a limit is imposed.
6. `handle_query_response` was revised to omit set and get helper functions in favor of direct Simple API calls, allowing for error handling and improved readability/maintainability.
7. `emit_select_statement` was modified to note the number of columns returned from the query for later use in `handle_query_response` and `get_plan_row_description`, which build up row information from columns incrementally (see point 8).
8. `handle_query_response` and `get_plan_row_description` were modified to use `ydb_get_s` calls and the total number of rows to retrieve row information, instead of using `ydb_subscript_next_s` as done previously. This saves some unnecessary iteration, improving performance. Now, each `RowDescription` is build up by iterating over the columns returned by the given query, up to the total number recorded by `emit_select_statement` per point 7. The row return process is thus to iterate over each result row and, for each result row, iterate over each result column to populate `RowDescription` information.

These changes are tested in new TJC011 test case in test_jdbc_connection.bats.in. Manual testing was also done with SquirrelSQL to confirm that rows are successfully limited without errors.

Note that some of the logic added for these changes is untested, as the JDBC driver doesn't seem to provide a mechanism for retrieving rows in chunks through alternating `PortalSuspended`/`Execute` messages. Rather, it allows limits to be specified for a single `Execute` query, after which a `Sync` is sent, terminating the extended query exchange before any subsequent rows are retrieved. This is also the behavior of SquirrelSQL.